Initial implementation of Timer
